refactor tests to single script
